feat: move function covertCellValue and covertCellValues to core package
