+++ packages/sheets/src/facade/utils.ts
@@ -14,15 +14,10 @@
  * limitations under the License.
  */
 
-import type { CellValue, ICellData, IObjectMatrixPrimitiveType, IRange, IRangeWithCoord, Nullable, Worksheet } from '@univerjs/core';
+import type { IRange, IRangeWithCoord, Worksheet } from '@univerjs/core';
 import {
     HorizontalAlign,
-    isCellV,
-    isFormulaString,
-    isICellData,
-    ObjectMatrix,
     RANGE_TYPE,
-    Tools,
     VerticalAlign,
 } from '@univerjs/core';
 
@@ -100,67 +95,6 @@ export function transformCoreVerticalAlignment(value: VerticalAlign): FVerticalA
         default:
             return 'general';
     }
-}
-
-/**
- * Covert cell value to cell data.
- * @param {CellValue | ICellData} value - The cell value.
- * @returns {ICellData} The cell data.
- */
-export function covertCellValue(value: CellValue | ICellData): ICellData {
-    if (isFormulaString(value)) {
-        return {
-            f: value as string,
-            v: null,
-            p: null,
-        };
-    }
-    if (isCellV(value)) {
-        return {
-            v: value as Nullable<CellValue>,
-            p: null,
-            f: null,
-        };
-    }
-    if (isICellData(value)) {
-        return value;
-    }
-
-    // maybe {}
-    return value as ICellData;
-}
-
-/**
- * Covert cell value array or matrix to cell data.
- * @param {CellValue[][] | IObjectMatrixPrimitiveType<CellValue> | ICellData[][] | IObjectMatrixPrimitiveType<ICellData>} value - The cell value array or matrix.
- * @param {IRange} range - The range.
- * @returns {IObjectMatrixPrimitiveType<ICellData>} The cell data matrix.
- */
-export function covertCellValues(
-    value:
-        | CellValue[][]
-        | IObjectMatrixPrimitiveType<CellValue>
-        | ICellData[][]
-        | IObjectMatrixPrimitiveType<ICellData>,
-    range: IRange
-): IObjectMatrixPrimitiveType<ICellData> {
-    const cellValue = new ObjectMatrix<ICellData>();
-    const { startRow, startColumn, endRow, endColumn } = range;
-
-    if (Tools.isArray(value)) {
-        for (let r = 0; r <= endRow - startRow; r++) {
-            for (let c = 0; c <= endColumn - startColumn; c++) {
-                cellValue.setValue(r + startRow, c + startColumn, covertCellValue(value[r][c]));
-            }
-        }
-    } else {
-        const valueMatrix = new ObjectMatrix(value as IObjectMatrixPrimitiveType<ICellData | CellValue>);
-        valueMatrix.forValue((r, c, v) => {
-            cellValue.setValue(r, c, covertCellValue(v));
-        });
-    }
-
-    return cellValue.getMatrix();
 }
 
 /**
